+++ lib/src/ast/sass/statement/content_block.dart
@@ -23,7 +23,6 @@ final class ContentBlock extends CallableDeclaration {
   T accept<T>(StatementVisitor<T> visitor) => visitor.visitContentBlock(this);
 
   @override
-  String toString() =>
-      (parameters.isEmpty ? "" : " using ($parameters)") +
-      " {${children.join(' ')}}";
+  String toString() => "${parameters.isEmpty ? "" : " using ($parameters)"} "
+      "{${children.join(' ')}}";
 }