+++ lib/src/js/legacy.dart
@@ -477,7 +477,7 @@ RenderResult _newRenderResult(
   if (_enableSourceMaps(options)) {
     var sourceMapOption = options.sourceMap;
     var sourceMapPath =
-        sourceMapOption is String ? sourceMapOption : options.outFile! + '.map';
+        sourceMapOption is String ? sourceMapOption : '${options.outFile!}.map';
     var sourceMapDir = p.dirname(sourceMapPath);
 
     var sourceMap = result.sourceMap!;