+++ lib/src/logger/stderr.dart
@@ -47,7 +47,7 @@ final class StderrLogger extends LoggerWithDeprecationType {
       // probably duplicated in the trace, so we just use it for highlighting.
       result.writeln(': $message\n\n${span.highlight(color: color)}');
     } else {
-      result.writeln(' on ${span.message("\n" + message, color: color)}');
+      result.writeln(' on ${span.message("\n$message", color: color)}');
     }
 
     if (trace != null) result.writeln(indent(trace.toString().trimRight(), 4));