+++ lib/src/visitor/async_evaluate.dart
@@ -1493,12 +1493,11 @@ final class _EvaluateVisitor
     for (var complex in styleRule.originalSelector.components) {
       if (!complex.isBogus) continue;
       _warn(
-        'The selector "${complex.toString().trim()}" is invalid CSS and ' +
-            (complex.isUseless ? "can't" : "shouldn't") +
-            ' be an extender.\n'
-                'This will be an error in Dart Sass 2.0.0.\n'
-                '\n'
-                'More info: https://sass-lang.com/d/bogus-combinators',
+        'The selector "${complex.toString().trim()}" is invalid CSS and '
+        '${complex.isUseless ? "can't" : "shouldn't"} be an extender.\n'
+        'This will be an error in Dart Sass 2.0.0.\n'
+        '\n'
+        'More info: https://sass-lang.com/d/bogus-combinators',
         MultiSpan(complex.span.trimRight(), 'invalid selector', {
           node.span: '@extend rule',
         }),
@@ -2527,22 +2526,21 @@ final class _EvaluateVisitor
             );
           }
         } else {
+          var willBeOmitted = complex.isBogusOtherThanLeadingCombinator
+              ? ' It will be omitted from the generated CSS.'
+              : '';
+          var suggestion = rule.children.every((child) => child is CssComment)
+              ? '\n(try converting to a //-style comment)'
+              : '';
           _warn(
             'The selector "${complex.toString().trim()}" is only valid for '
-                    "nesting and shouldn't\n"
-                    'have children other than style rules.' +
-                (complex.isBogusOtherThanLeadingCombinator
-                    ? ' It will be omitted from the generated CSS.'
-                    : '') +
-                '\n'
-                    'This will be an error in Dart Sass 2.0.0.\n'
-                    '\n'
-                    'More info: https://sass-lang.com/d/bogus-combinators',
+            "nesting and shouldn't\n"
+            'have children other than style rules.$willBeOmitted\n'
+            'This will be an error in Dart Sass 2.0.0.\n'
+            '\n'
+            'More info: https://sass-lang.com/d/bogus-combinators',
             MultiSpan(complex.span.trimRight(), 'invalid selector', {
-              rule.children.first.span: "this is not a style rule" +
-                  (rule.children.every((child) => child is CssComment)
-                      ? '\n(try converting to a //-style comment)'
-                      : ''),
+              rule.children.first.span: "this is not a style rule$suggestion",
             }),
             Deprecation.bogusCombinators,
           );
@@ -2948,9 +2946,7 @@ final class _EvaluateVisitor
 
     if (results == null) return sassNull;
     return SassString(
-        'if(' +
-            results.map((pair) => '${pair.$1}: ${pair.$2}').join('; ') +
-            ')',
+        'if(${results.map((pair) => '${pair.$1}: ${pair.$2}').join('; ')})',
         quotes: false);
   }
 
@@ -3003,10 +2999,8 @@ final class _EvaluateVisitor
   @override
   Future<Object /* String | bool */ > visitIfConditionFunction(
           IfConditionFunction node) async =>
-      (await _performInterpolation(node.name)) +
-      '(' +
-      (await _performInterpolation(node.arguments)) +
-      ')';
+      '${await _performInterpolation(node.name)}'
+      '(${await _performInterpolation(node.arguments)})';
 
   @override
   Future<Object /* String | bool */ > visitIfConditionSass(
@@ -3265,13 +3259,11 @@ final class _EvaluateVisitor
           node.arguments.positional.length > maxArgs) {
         throw _exception(
           "Only $maxArgs ${pluralize('argument', maxArgs)} allowed, but "
-                  "${node.arguments.positional.length} " +
-              pluralize(
-                'was',
-                node.arguments.positional.length,
-                plural: 'were',
-              ) +
-              " passed.",
+          "${node.arguments.positional.length} ${pluralize(
+            'was',
+            node.arguments.positional.length,
+            plural: 'were',
+          )} passed.",
           node.span,
         );
       }